+++ app/src/main/java/org/stepic/droid/di/mainscreen/MainScreenComponent.kt
@@ -2,10 +2,15 @@ package org.stepic.droid.di.mainscreen
 
 import dagger.Subcomponent
 import org.stepic.droid.ui.activities.MainFeedActivity
+import org.stepik.android.view.injection.user.UserDataModule
 import org.stepik.android.view.injection.user_profile.UserProfileDataModule
 
 @MainScreenScope
-@Subcomponent(modules = [MainScreenModule::class, UserProfileDataModule::class])
+@Subcomponent(modules = [
+    MainScreenModule::class,
+    UserDataModule::class,
+    UserProfileDataModule::class
+])
 interface MainScreenComponent {
 
     @Subcomponent.Builder
