+++ app/src/main/java/org/stepik/android/presentation/profile/ProfileView.kt
@@ -1,28 +1,18 @@
 package org.stepik.android.presentation.profile
 
-import org.stepic.droid.model.AchievementFlatItem
-import org.stepic.droid.model.UserViewModel
+import org.stepik.android.model.user.User
+
 
 interface ProfileView {
     sealed class State {
         object Idle : State()
-        object NetworkError : State()
-        object UserNotFoundError : State()
-        object NeedAuthError : State()
         object Loading : State()
-        data class ProfileLoaded(val userLocalViewModel: UserViewModel) : State()
+
+        class Content(val user: User) : State()
+        object Empty : State()
+        object NetworkError : State()
     }
 
     fun setState(state: State)
-
-    // Achievements related
-    fun showAchievements(achievements: List<AchievementFlatItem>)
-    fun onAchievementsLoadingError()
-    fun onAchievementsLoading()
-
-    // Streak related
-    fun showNotificationEnabledState(notificationEnabled: Boolean, notificationTimeValue: String)
-    fun hideNotificationTime(needHide: Boolean)
-    fun setNewTimeInterval(timePresentationString: String)
-    fun onStreaksLoaded(currentStreak: Int, maxStreak: Int, haveSolvedToday: Boolean)
+    fun showNetworkError()
 }
