+++ app/src/main/java/org/stepik/android/view/profile/ui/fragment/ProfileFragment.kt
@@ -7,7 +7,10 @@ import android.view.MenuInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.text.bold
+import androidx.core.text.buildSpannedString
 import androidx.core.view.ViewCompat
+import androidx.core.view.isVisible
 import androidx.core.widget.NestedScrollView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
@@ -17,16 +20,21 @@ import kotlinx.android.synthetic.main.fragment_profile.*
 import kotlinx.android.synthetic.main.header_profile.*
 import org.stepic.droid.R
 import org.stepic.droid.base.App
+import org.stepic.droid.ui.util.snackbar
 import org.stepik.android.presentation.profile.ProfilePresenter
 import org.stepik.android.presentation.profile.ProfileView
 import org.stepik.android.view.ui.delegate.ViewStateDelegate
+import ru.nobird.android.view.base.ui.extension.argument
 import javax.inject.Inject
 import kotlin.math.min
 
-class ProfileFragment : Fragment() {
+class ProfileFragment : Fragment(), ProfileView {
     companion object {
-        fun newInstance(): Fragment =
+        fun newInstance(userId: Long): Fragment =
             ProfileFragment()
+                .apply {
+                    this.userId = userId
+                }
     }
 
     @Inject
@@ -34,6 +42,7 @@ class ProfileFragment : Fragment() {
 
     private lateinit var profilePresenter: ProfilePresenter
 
+    private var userId by argument<Long>()
 
     private lateinit var viewStateDelegate: ViewStateDelegate<ProfileView.State>
 
@@ -45,6 +54,7 @@ class ProfileFragment : Fragment() {
         profilePresenter = ViewModelProviders
             .of(this, viewModelFactory)
             .get(ProfilePresenter::class.java)
+        profilePresenter.onData(userId)
     }
 
     override fun onCreateView(
@@ -65,17 +75,6 @@ class ProfileFragment : Fragment() {
                 setDisplayShowTitleEnabled(false)
             }
 
-        profileName.text = "Konstantin Konstantin"
-        profileBio.text = "Saint Petersburg State University, Bioinformatics Institute, VK, Saint Petersburg State University, Bioinformatics Institute, VK, Saint Petersburg State University, Bioinformatics Institute, VK"
-
-        toolbarTitle.text = profileName.text
-        toolbarTitle.translationY = 1000f
-
-        Glide
-            .with(this)
-            .load("https://i.pinimg.com/originals/a2/de/39/a2de3954697c636276192afea0a6f661.jpg")
-            .into(profileImage)
-
         ViewCompat.setElevation(header, resources.getDimension(R.dimen.profile_header_elevation))
 
         scrollContainer.setOnScrollChangeListener { _: NestedScrollView, _: Int, scrollY: Int, _: Int, _: Int ->
@@ -84,6 +83,22 @@ class ProfileFragment : Fragment() {
             val scroll = min(toolbar.height, scrollY)
             toolbarTitle.translationY = toolbar.height.toFloat() - scroll
         }
+
+        profileKnowledgeRank.isVisible = false
+        profileReputationRank.isVisible = false
+
+//        profileKnowledgeRank.text =
+//            buildSpannedString {
+//                append("Знания ")
+//                bold { append("2132") }
+//                append(" (25й по знаниям)")
+//            }
+//        profileReputationRank.text =
+//            buildSpannedString {
+//                append("Репутация ")
+//                bold { append("132") }
+//                append(" (250й по репутации)")
+//            }
     }
 
     private fun injectComponent() {
@@ -99,18 +114,43 @@ class ProfileFragment : Fragment() {
 
     override fun onStart() {
         super.onStart()
-
-//        profilePresenter.attachView(this)
+        profilePresenter.attachView(this)
     }
 
     override fun onStop() {
-//        profilePresenter.detachView(this)
+        profilePresenter.detachView(this)
         super.onStop()
     }
 
-//    override fun setState(state: ProfileView.State) {
-//        when (state) {
-//
-//        }
-//    }
+    override fun setState(state: ProfileView.State) {
+        viewStateDelegate.switchState(state)
+
+        when (state) {
+            is ProfileView.State.Content -> {
+                Glide
+                    .with(this)
+                    .load(state.user.avatar)
+                    .placeholder(R.drawable.general_placeholder)
+                    .into(profileImage)
+
+                profileName.text = state.user.fullName
+                profileBio.text = state.user.shortBio
+                profileBio.isVisible = !state.user.shortBio.isNullOrBlank()
+
+                toolbarTitle.text = state.user.fullName
+                toolbarTitle.translationY = 1000f
+            }
+
+            ProfileView.State.Loading,
+            ProfileView.State.Empty,
+            ProfileView.State.NetworkError -> {
+                toolbarTitle.setText(R.string.profile_title)
+                toolbarTitle.translationY = 0f
+            }
+        }
+    }
+
+    override fun showNetworkError() {
+        view?.snackbar(messageRes = R.string.connectionProblems)
+    }
 }
