+++ app/src/main/res/drawable/ic_profile_knowledge.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+
+    <size android:width="12dp" android:height="12dp" />
+    <solid android:color="@color/green01" />
+
+</shape>
